backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
